+++ marketplace/accounts/permissions.py
@@ -1,4 +1,5 @@
 from rest_framework import permissions
+
 from django.contrib.auth.models import AnonymousUser
 from django.conf import settings