+++ marketplace/webhooks/facebook/views.py
@@ -18,4 +18,11 @@ class FacebookWebhook(APIView):  # pragma: no cover
             name="update_templates_by_webhook",
             kwargs={"webhook_data": request.data},
         )
+
+        # Also send the task to update account status since multiple entries can exist in a single webhook
+        celery_app.send_task(
+            name="update_account_info_by_webhook",
+            kwargs={"webhook_data": request.data},
+        )
+
         return Response(status=status.HTTP_200_OK)